resolution change without restart
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@239 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ game_type_select.cpp
@@ -89,7 +89,7 @@ static void GameSelectInit (void) {
 	scope = 1;
 	
 	ResetWidgets ();
-	double top = (double)(param.y_resolution * 0.4);
+	int top = (int)(param.y_resolution * 0.4);
 	double dist = AutoDistance ();
 	AddTextButton (Trans.Text(1), CENTER, top, 0, FIT);
 	AddTextButton (Trans.Text(2), CENTER, top + dist, 1, FIT);
